django_server/requirements.txt
channels==4.2.0
Django==5.1.6
djangorestframework==3.15.2
